[AMDGPU] Avoid unnecessary graph visits during WQM marking
(Cherry-picked from upstream)

Avoid revisiting nodes with the same set of defined lanes by
using a unified visited set which integrates lanes into the key.
This retains the intent of the original code by still revisiting
a subgraph if a different set of lanes is defined and hence
marking might progress differently.

Note: default size of the visited set has been confirmed to
cover >99% of invocations in large array of test shaders.

Reviewed By: piotr

Differential Revision: https://reviews.llvm.org/D98772

Change-Id: I0f5e2246e26f78106557cddeb433d41b04784b23