[ADT] Add all_equal predicate
`llvm::all_equal` checks if all values in the given range are equal,
i.e., there are no two elements that not equal.
Similar to `llvm::all_of`, it returns `true` when the range is emtpy.

`llvm::all_equal` is intended to supersede `llvm::is_splat`, which will
be deprecated and removed in future patches. See the discussion thread
for more details:
https://discourse.llvm.org/t/adt-is-splat-and-empty-ranges/64692.