[WIP][STLExtras] Allow `llvm::enumerate` to enumerate over multiple ranges
This does not work by a mere composition of `enumerate` and `zip_equal`,
because C++17 does not allow for recursive expansion of structured
bindings.

This implementation uses `zippy` to manage the iteratees and adds the
stream of indices as the first zipped range. Because we have an upfront
assertion that all input ranges are of the same length, we only need to
check if the second range has ended during iteration.

I marked this as 'Work In Progress' because I'm not sure if there's a
simpler/cleaner implementation. I'd really appreciate some suggestions.

Differential Revision: https://reviews.llvm.org/D144503